requirements.txt
amqp==5.3.1
asgiref==3.10.0
bcrypt==3.2.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
concurrent-log-handler==0.9.28
cron_descriptor==2.0.6
Django==5.2.7
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-filter==25.2
django-js-asset==3.1.2
django-mptt==0.18.0
django-redis==6.0.0
django-restframework==0.0.1
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
et_xmlfile==2.0.0
greenlet==3.2.4
idna==3.11
inflection==0.5.1
jsonpath==0.82.2
kombu==5.5.4
mysqlclient==2.2.7
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==12.0.0
portalocker==3.2.0
prompt_toolkit==3.0.52
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
qrcode==8.2
redis==7.0.0
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.44
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
